latest

+++ app/modules/expense/expense.model.js
@@ -24,10 +24,10 @@ module.exports = (sequelize, DataTypes) => {
       },
 
       amount: {
-        type: DataTypes.FLOAT,
+        type: DataTypes.DECIMAL(15, 2),
         allowNull: false,
         validate: {
-          notEmpty: true, // Ensure name is not empty
+          notEmpty: true,
         },
       },
     },
