+++ app/modules/meta/meta.model.js
@@ -19,10 +19,10 @@ module.exports = (sequelize, DataTypes) => {
         },
       },
       amount: {
-        type: DataTypes.FLOAT,
+        type: DataTypes.DECIMAL(15, 2),
         allowNull: false,
         validate: {
-          notEmpty: true, // Ensure name is not empty
+          notEmpty: true,
         },
       },
       note: {
