+++ app/modules/pettyCash/pettyCash.model.js
@@ -25,10 +25,10 @@ module.exports = (sequelize, DataTypes) => {
         allowNull: true,
       },
       amount: {
-        type: DataTypes.FLOAT,
+        type: DataTypes.DECIMAL(15, 2),
         allowNull: false,
         validate: {
-          notEmpty: true, // Ensure name is not empty
+          notEmpty: true,
         },
       },
       note: {
